+++ CHANGES.rst
@@ -12,6 +12,7 @@ mio 0.0.6.dev
 - Refactored all of the context state management code (*stopStatus*) and exposed it to the end user.
 
   - This means we can now write flow based constructs such as loops directly in mio.
+- Fixed a minor bug in the parser where ``not(0) ifTrue(print("foo"))`` would parse as ``not(0, ifTrue(print("foo")))``
 
 
 mio 0.0.5 (2013-10-29)